Moved functions outside of main into logic.dart
